requirements.txt
markdownify
smolagents==1.13.0
requests
duckduckgo_search
pandas
pytz
PyYAML
Pillow
gradio
python-dotenv
